MetabolicEP.metabolicEP — Method

res=metabolicEP(S,b,lb,ub,...)

The output in res is of type EPout: there are several fields:

μ::Vector: A parameter linked to the mean of the posterior probability

σ::Vector: A parameter linked to the std of the posterior probability

av::Vector: The mean posterior probability

va::Vector: The variance of the posterior probability

sol::EPFields: The internal field status. From this value we can restart the sampling from a specific state.

status::Symbol: either $:converged$ or $:unconverged$.

Input (required)

S: MxN matrix (either sparse or dense) please note that if you input a dense version, the algorithm is slighlty more efficient. Dense matrices can be create from sparse ones with Matrix(S).

b: a vector of M intakes/uptakes

lb: a vector of lengh N of lower bounds.

ub: a vector of lengh N of upper bounds.

Input (optional arguments).

beta (inverse temperature::$Real$): default 10^7; the zero temperature algorithm is run setting $beta=Inf$.

verbose ($true$ or $false$): default $true$

damp (∈ (0,1) newfield = damp * oldfield + (1-damp)* newfield): default 0.9

epsconv (convergence criterion): default 1e-6

maxiter (maximum number of iterations): default 2000

maxvar (threshold on maximum variance): default 1e50

minvar (threshold on minimum variance): default 1e-50

solution (start from solution. Is of type $EPout$): default: $nothing$

expval (fix to posterior probability of mean and/or variance to values): default $nothing$. expval can be either at $Tuple{Float64,Float64,Int}$ or a $Vector{Tuple{Float64,Float64,Int}}$. Values can be fixed as$expval=(0.2,0.4,4)$ meaning that for flux index 4 the mean is set to 0.2 and the variance to 0.4. Fixing more values $expval=[(0.2, 0.3, 4), (0.4, nothing, 5)]$: in this case, we fix the posterior of flux 4 to 0.2 (mean) and 0.3 (variance), while for flux 5 we fix the mean to 0.4 and we keep the variance free.

MetabolicEP Documentation

Missing docstring.

Missing docstring for MetabolicEP. Check Documenter's build log for details.